+++ authentication/src/main/kotlin/com/uber/sdk2/auth/internal/sso/UniversalSsoLink.kt
@@ -23,13 +23,13 @@ import androidx.annotation.VisibleForTesting
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatActivity.RESULT_CANCELED
 import androidx.appcompat.app.AppCompatActivity.RESULT_OK
-import com.uber.sdk2.auth.api.AppDiscovering
-import com.uber.sdk2.auth.api.exception.AuthException
-import com.uber.sdk2.auth.api.request.AuthContext
-import com.uber.sdk2.auth.api.request.AuthDestination
-import com.uber.sdk2.auth.api.request.SsoConfig
-import com.uber.sdk2.auth.api.sso.CustomTabsLauncher
-import com.uber.sdk2.auth.api.sso.SsoLink
+import com.uber.sdk2.auth.AppDiscovering
+import com.uber.sdk2.auth.exception.AuthException
+import com.uber.sdk2.auth.request.AuthContext
+import com.uber.sdk2.auth.request.AuthDestination
+import com.uber.sdk2.auth.request.SsoConfig
+import com.uber.sdk2.auth.sso.CustomTabsLauncher
+import com.uber.sdk2.auth.sso.SsoLink
 import com.uber.sdk2.core.config.UriConfig
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Dispatchers
