+++ authentication/src/main/kotlin/com/uber/sdk2/auth/request/AuthDestination.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.uber.sdk2.auth.api.request
+package com.uber.sdk2.auth.request
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
