+++ authentication/src/test/kotlin/com/uber/sdk2/auth/internal/AuthProviderTest.kt
@@ -16,23 +16,23 @@
 package com.uber.sdk2.auth.internal
 
 import androidx.appcompat.app.AppCompatActivity
+import com.uber.sdk2.auth.PKCEGenerator
 import com.uber.sdk2.auth.RobolectricTestBase
-import com.uber.sdk2.auth.api.PKCEGenerator
-import com.uber.sdk2.auth.api.exception.AuthException
-import com.uber.sdk2.auth.api.request.AuthContext
-import com.uber.sdk2.auth.api.request.AuthDestination
-import com.uber.sdk2.auth.api.request.AuthType
-import com.uber.sdk2.auth.api.request.CrossApp
-import com.uber.sdk2.auth.api.request.PrefillInfo
-import com.uber.sdk2.auth.api.response.AuthResult
-import com.uber.sdk2.auth.api.response.PARResponse
-import com.uber.sdk2.auth.api.response.UberToken
-import com.uber.sdk2.auth.api.sso.SsoLink
+import com.uber.sdk2.auth.exception.AuthException
 import com.uber.sdk2.auth.internal.service.AuthService
 import com.uber.sdk2.auth.internal.shadow.ShadowSsoConfigProvider
 import com.uber.sdk2.auth.internal.shadow.ShadowSsoLinkFactory
 import com.uber.sdk2.auth.internal.sso.SsoLinkFactory
 import com.uber.sdk2.auth.internal.utils.Base64Util
+import com.uber.sdk2.auth.request.AuthContext
+import com.uber.sdk2.auth.request.AuthDestination
+import com.uber.sdk2.auth.request.AuthType
+import com.uber.sdk2.auth.request.CrossApp
+import com.uber.sdk2.auth.request.PrefillInfo
+import com.uber.sdk2.auth.response.AuthResult
+import com.uber.sdk2.auth.response.PARResponse
+import com.uber.sdk2.auth.response.UberToken
+import com.uber.sdk2.auth.sso.SsoLink
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
